requirements.txt
# DeepSeek API (OpenAI-compatible)
openai>=1.0.0

# Yandex Calendar (CalDAV)
caldav>=1.3.0
icalendar>=5.0.0

# Telegram (MTProto)
telethon>=1.34.0

# Scheduling
apscheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging
python-json-logger>=2.0.0
